Change the type of the `points` field in the `User` struct from `u64` to `u32`
Change the type of the `points_to_add` variable from `u64` to `u32` in `purchase_points` function

Ensure that the `updated_points` variable is of type `u32` in `main.rs`

Add `extract_email_from_jwt` function to the `auth` module

Add a migration file to change the column type from `BIGINT` to `INTEGER`

+++ backend/src/db.rs
@@ -47,8 +47,8 @@ pub async fn login_user(client: &Client, email: &str, password: &str) -> Result<
     }
 }
 
-pub async fn purchase_points(client: &Client, email: &str, amount: f64) -> Result<u64, MyError> {
-    let points_to_add = (amount * 100.0) as u64;
+pub async fn purchase_points(client: &Client, email: &str, amount: f64) -> Result<u32, MyError> {
+    let points_to_add = (amount * 100.0) as u32;
 
     let stmt = include_str!("sql/get_user_by_email.sql");
     let stmt = stmt.replace("$table_fields", &User::sql_table_fields());
