+++ backend/src/models.rs
@@ -8,5 +8,5 @@ pub struct User {
     pub email: String,
     pub username: String,
     pub password: String,
-    pub points: u64,
+    pub points: u32,
 }
